chore: use the shared trade input body
